refactor: Utilize formatDurationText utility for duration formatting in TicketsPage and server methods
- Replaced manual duration formatting logic with the new formatDurationText utility function in TicketsPage, ClockEventHelpers, and related server methods.
- Improved code maintainability and consistency in duration representation across the application.
