+++ client/components/tickets/TicketsPage.js
@@ -2,7 +2,7 @@ import { Template } from 'meteor/templating';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { Teams, Tickets, ClockEvents } from '../../../collections.js';
 import { currentTime } from '../layout/MainLayout.js';
-import { formatTime, calculateTotalTime } from '../../utils/TimeUtils.js';
+import { formatTime, calculateTotalTime, formatDurationText } from '../../utils/TimeUtils.js';
 import { extractUrlTitle } from '../../utils/UrlUtils.js';
 import { getUserTeams } from '../../utils/UserTeamUtils.js';
 
@@ -139,16 +139,9 @@ Template.tickets.onCreated(function () {
         if (sessionDurationSeconds >= TEN_HOURS_SECONDS) {
           this.autoClockOutTriggered.set(true);
           
-          // Calculate total duration for notification (same format as server)
+          // Calculate total duration for notification (using utility function)
           const totalSeconds = (activeSession.accumulatedTime || 0) + sessionDurationSeconds;
-          const hours = Math.floor(totalSeconds / 3600);
-          const minutes = Math.floor((totalSeconds % 3600) / 60);
-          const seconds = totalSeconds % 60;
-          const parts = [];
-          if (hours > 0) parts.push(`${hours}h`);
-          if (minutes > 0) parts.push(`${minutes}m`);
-          if (seconds > 0 || parts.length === 0) parts.push(`${seconds}s`);
-          const durationText = parts.join(' ');
+          const durationText = formatDurationText(totalSeconds);
           
           // Get team name for notification
           const team = Teams.findOne(teamId);
