+++ client/utils/TimeUtils.js
@@ -29,6 +29,30 @@ export const formatTime = (time) => {
 };
 
 /**
+ * Formats a time duration in seconds into "Xh Ym Zs" format
+ * @param {number} totalSeconds - Time in seconds
+ * @returns {string} Formatted duration string (e.g., "10h 0m 0s" or "5m 30s")
+ * @example
+ * formatDurationText(3661) // Returns "1h 1m 1s"
+ * formatDurationText(36000) // Returns "10h"
+ * formatDurationText(90) // Returns "1m 30s"
+ * formatDurationText(0) // Returns "0s"
+ */
+export const formatDurationText = (totalSeconds) => {
+    if (typeof totalSeconds !== 'number' || isNaN(totalSeconds) || totalSeconds < 0) {
+        return '0s';
+    }
+    const hours = Math.floor(totalSeconds / 3600);
+    const minutes = Math.floor((totalSeconds % 3600) / 60);
+    const seconds = totalSeconds % 60;
+    const parts = [];
+    if (hours > 0) parts.push(`${hours}h`);
+    if (minutes > 0) parts.push(`${minutes}m`);
+    if (seconds > 0 || parts.length === 0) parts.push(`${seconds}s`);
+    return parts.join(' ');
+};
+
+/**
  * Formats a timestamp into a locale-specific date string
  * @param {number} timestamp - Unix timestamp in milliseconds
  * @returns {string} Formatted date string
